Fix a performance issue introduced in 0.2.0
I had extracted a method named _rails_cb_log to reduce duplication
but that made my test suite at work go from 200s to 370s, or 85%
slower!

The performance regression could be directly due to the single additional
method dispatch, or it could be indirect, e.g. a larger "scope" for the
lambda.